requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
pydantic==2.10.2
python-dotenv==1.0.1
firebase-admin==6.6.0
